requirements.txt
PyQt6 
reportlab 
python-escpos 
pandas 
openpyxl 
Babel 
cryptography 
qtawesome 
bcrypt 
python-dotenv
pyinstaller
opencv-python-headless
numpy
psutil
google-api-python-client 
google-auth-oauthlib 
google-auth-httplib2
google-auth
supabase
python-dateutil
requests
py-machineid
qtpy
pillow
pyarmor
python-barcode
qrcode